+++ src/main.cpp
@@ -67,7 +67,7 @@ class $modify(SpeedrunPlayLayer, PlayLayer) {
                         ->setAutoScale(false)
                         ->setGap(3.75f);
 
-                    auto opacity = static_cast<int>(srt->getSettingValue<int64_t>("mobile-btns-opacity"));
+                    auto opacity = srt->getSettingValue<int64_t>("mobile-btns-opacity");
 
                     // menu for mobile controls
                     m_fields->m_mobileMenu = CCMenu::create();
